Moved executable from cmd to pks, matching 'agent' structure
update package names, Makefile and Dockerfile to match new structure
